% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adb_data.R
\name{adb_data}
\alias{adb_data}
\title{Load the data from an areal database}
\usage{
adb_data(
  root,
  ...,
  variables = NULL,
  merge = FALSE,
  outPath = NULL,
  outType = NULL
)
}
\arguments{
\item{root}{[\code{character(1)}]\cr path to the root directory that contains
an areal database.}

\item{...}{[\code{character(1)}]\cr named argument, typically of identifying
variables, to filter by.}

\item{variables}{[\code{character(.)}]\cr columns, typically observed
variables, to select.}

\item{merge}{[\code{logical(1)}]\cr whether or not to merge tables and
geometries for the specified query. At the default value (FALSE), tables
and geometries are provided as separate files.}

\item{outPath}{description}

\item{outType}{[\code{character(1)}]\cr the output file-type, see
\code{\link{st_drivers}} for a list. If a file-type supports layers, they
are stored in the same file, otherwise the different layers are provided
separately. For an R-based workflow, \code{"rds"} could be an efficient
option.}
}
\value{
if \code{outPath} and \code{outType} is given, there is no return
  value (instead, the resulting object is stored in \code{outPath}),
  otherwise ...
}
\description{
Load the data from an areal database
}
