% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adb_init.R
\name{adb_init}
\alias{adb_init}
\title{Initiate an areal database}
\usage{
adb_init(root, version, author, licence, level)
}
\arguments{
\item{root}{\code{\link[=character]{character(1)}}\cr path to the root directory that
contains or shall contain an areal database.}

\item{version}{\code{\link[=character]{character(1)}}\cr version identifier for this
areal database.}

\item{author}{\code{\link[=character]{character(1)}}\cr authors that contributed to
building this areal database. Should be a list with items \code{"cre"}
(creator), \code{"aut"} (authors) and \code{"ctb"} (contributors).}

\item{licence}{\code{\link[=character]{character(1)}}\cr licence (link) for this areal
database.}

\item{level}{\code{\link[=character]{character(1)}}\cr the label of the class in the
gazetteer at which this database operates (e.g. \code{"ADM0"} for a
national database). Stage 3 outputs are split by units of this class.}
}
\value{
No return value, called for the side effect of creating the
directory structure of the new areal database and tables that contain the
database metadata.
}
\description{
Initiate a geospatial database or register a database that exists at the root
path.
}
\details{
This is the first function that is run in a project, as it initiates
the areal database by creating the default sub-directories and initial
inventory tables. Vocabularies (gazetteer, ontologies) are added later via
\code{\link{regVocabulary}} and \code{\link{normVocabulary}}.
}
\examples{
adb_init(root = paste0(tempdir(), "/newDB"),
         version = "1.0.0", licence = "CC-BY-0.4",
         author = list(cre = "Jane Doe", aut = "John Doe", ctb = "Jamie Roe"),
         level = "ADM0")
}
