% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adb_inventory.R
\name{adb_inventory}
\alias{adb_inventory}
\title{Load the inventory of the currently active areal database}
\usage{
adb_inventory(type = NULL)
}
\arguments{
\item{type}{\code{\link[=character]{character(1)}}\cr the inventory sub-table to load,
either \code{"dataseries"}, \code{"tables"}, \code{"geometries"} or
\code{"references"}.}
}
\value{
returns the table selected in \code{type}
}
\description{
Load the inventory of the currently active areal database
}
