% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adb_metadata.R
\name{adb_metadata}
\alias{adb_metadata}
\title{Summarise database metadata}
\usage{
adb_metadata()
}
\value{
a list with elements \code{$nations}, \code{$variables},
\code{$concepts}, \code{$years}.
}
\description{
Returns a summary of what is currently in the database: nations covered,
variables available, thematic concepts, and time range.
}
