% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adb_querry.R
\name{adb_querry}
\alias{adb_querry}
\title{Extract database contents}
\usage{
adb_querry(
  territory = NULL,
  concept = NULL,
  variable = NULL,
  level = NULL,
  year = NULL
)
}
\arguments{
\item{territory}{combination of column name in the ontology and value to
filter that column by to build a tree of the territories nested into it;
see \code{\link[ontologics]{make_tree}}.}

\item{concept}{description}

\item{variable}{description}

\item{level}{description}

\item{year}{description}
}
\value{
returns ...
}
\description{
Extract database contents
}
