% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adb_reset.R
\name{adb_reset}
\alias{adb_reset}
\title{Reset an areal database to its unfilled state}
\usage{
adb_reset(what = "all")
}
\arguments{
\item{what}{\code{\link[=character]{character(.)}}\cr what to reset; one or more of
\code{"vocabularies"}, \code{"schemas"}, \code{"tables"},
\code{"geometries"}, \code{"inventory"}, or \code{"all"} (the default).}
}
\value{
no return value, called for its side effect of reorganising an areal
database into a state where no reg* or norm* functions have been run
}
\description{
Reset an areal database to its unfilled state
}
