% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adb_schemas.R
\name{adb_schemas}
\alias{adb_schemas}
\title{Load the schemas of the currently active areal database}
\usage{
adb_schemas(pattern = NULL)
}
\arguments{
\item{pattern}{\code{\link[=character]{character(1)}}\cr an optional regular expression.
Only schema names which match the regular expression will be processed.}
}
\value{
returns a list of schema descriptions
}
\description{
Load the schemas of the currently active areal database
}
