% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{countries}
\alias{countries}
\title{A table with country level metadata}
\format{The object of class tibble has 7 columns and 249 rows. It lists for
  each country the \code{name}, \code{iso_a2} and \code{iso-a3}-code, whether
  the country is an \code{un_member}, on which \code{continent} the country
  is and of which \code{region} and \code{subregion} that is part.}
\usage{
countries
}
\description{
A table with country level metadata
}
\keyword{datasets}
