% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{delayedRead}
\alias{delayedRead}
\title{Open file when it's not open}
\usage{
delayedRead(file, fun = NULL, ...)
}
\arguments{
\item{file}{a file to store with a delay.}

\item{fun}{the function to use.}

\item{...}{arguments to the \emph{write} functions.}
}
\description{
Open file when it's not open
}
\details{
This function takes the \code{file} and only opens it, if it is not
currently already open. In the case it's already open, it prompts the user to
confirm when the file is closed (and processed by the other process) and then
opens it again.
}
