% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.getColTypes}
\alias{.getColTypes}
\title{Get the column types of a tibble}
\usage{
.getColTypes(input = NULL)
}
\arguments{
\item{input}{\link{data.frame}\cr table from which to get column
types.}
}
\description{
(internal function not for user interaction)
}
