% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.read_all_mappings}
\alias{.read_all_mappings}
\title{Read all mappings files for one vocabulary}
\usage{
.read_all_mappings(vName)
}
\arguments{
\item{vName}{\code{\link[=character]{character(1)}}\cr vocabulary name}
}
\value{
named list of tibbles, one per dataseries
}
\description{
Read all mappings files for one vocabulary
}
