% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.read_levels}
\alias{.read_levels}
\title{Read the levels table for a vocabulary}
\usage{
.read_levels(vName)
}
\arguments{
\item{vName}{\code{\link[=character]{character(1)}}\cr vocabulary name}
}
\value{
tibble with columns: level_id, label, parent_level
}
\description{
Read the levels table for a vocabulary
}
