% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.read_mappings}
\alias{.read_mappings}
\title{Read the mappings table for one dataseries against one vocabulary}
\usage{
.read_mappings(vName, dataseries)
}
\arguments{
\item{vName}{\code{\link[=character]{character(1)}}\cr vocabulary name}

\item{dataseries}{\code{\link[=character]{character(1)}}\cr dataseries name}
}
\value{
tibble with columns: source_label, source, canonical_id, note
or an empty tibble with those columns if the file doesn't exist
}
\description{
Read the mappings table for one dataseries against one vocabulary
}
