% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.read_terms}
\alias{.read_terms}
\title{Read the terms table for a vocabulary}
\usage{
.read_terms(vName)
}
\arguments{
\item{vName}{\code{\link[=character]{character(1)}}\cr vocabulary name}
}
\value{
tibble with columns: id, label, class, parent_id, description,
valid_from, valid_until, successor_id
}
\description{
Read the terms table for a vocabulary
}
