% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.updateOntology}
\alias{.updateOntology}
\title{Update an ontology}
\usage{
.updateOntology(
  table = NULL,
  threshold = NULL,
  dataseries = NULL,
  ontology = NULL
)
}
\arguments{
\item{table}{\code{\link[=character]{character(1)}}\cr a table that contains a match
column as the basis to update the gazetteer.}

\item{threshold}{\code{\link[=numeric]{numeric(1)}}\cr a threshold value above which
matches are updated in the gazetteer.}

\item{dataseries}{\code{\link[=character]{character(1)}}\cr the source dataseries of the
external concepts for which the gazetteer shall be updated.}

\item{ontology}{\link[ontologics:onto-class]{onto}\cr path where the ontology/gazetteer
is stored.}
}
\value{
called for its side-effect of updating a gazetteer
}
\description{
This function takes a table (spatial) and updates all territorial concepts in
the provided gazetteer.
}
