% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.write_terms}
\alias{.write_terms}
\title{Append new terms to \if{html}{\out{<vName>}}_terms.parquet, ignoring duplicates}
\usage{
.write_terms(vName, new_rows)
}
\arguments{
\item{vName}{\code{\link[=character]{character(1)}}\cr vocabulary name}

\item{new_rows}{tibble with columns matching the terms schema}
}
\description{
Append new terms to \if{html}{\out{<vName>}}_terms.parquet, ignoring duplicates
}
