% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geom_countries}
\alias{geom_countries}
\title{Simple feature geometries of all countries}
\format{An object of class sf (inherits from tbl_df, tbl, data.frame) with
  177 rows and 7 columns}
\usage{
geom_countries
}
\description{
Simple feature geometries of all countries
}
\keyword{datasets}
