% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_variable.R
\name{get_variable}
\alias{get_variable}
\title{Extract concepts in stage2 tables}
\usage{
get_variable(variable = NULL, dataseries = NULL)
}
\arguments{
\item{variable}{[\code{character(.)}]\cr name of the variable for which to
extract concepts.}

\item{dataseries}{[\code{integerish(1)}]\cr name of the dataseries from which
to extract the variables' concepts.}
}
\value{
The table provided in \code{input}, where the given variable is
  replaced by the column that is specified by the argument name.
}
\description{
This function matches the values of a variable with an index and returns the
specified IDs.
}
