% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_gazetteer.R
\name{match_gazetteer}
\alias{match_gazetteer}
\title{Match concepts in a table with a gazetteer}
\usage{
match_gazetteer(
  table = NULL,
  columns = NULL,
  dataseries = NULL,
  from_meta = FALSE
)
}
\arguments{
\item{table}{[\code{character(1)}]\cr a table that contains columns that should be harmonised by matching it with the gazetteer.}

\item{columns}{[\code{character(1)}]\cr the columns containing the concepts}

\item{dataseries}{[\code{character(1)}]\cr}

\item{from_meta}{[\code{logical(1)}]\cr}
}
\description{
Match concepts in a table with a gazetteer
}
