% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchCommodities.R
\name{matchCommodities}
\alias{matchCommodities}
\title{Determine the valid ID of commodities}
\usage{
matchCommodities(input = NULL, ..., keepOrig = FALSE)
}
\arguments{
\item{input}{[\code{character(.)}]\cr terms to be translated.}

\item{...}{[\code{data.frame(1)}]\cr define columns in \code{input}, where
the commodities are stored; see Details.}

\item{keepOrig}{[\code{logical(1)}]\cr to keep the original commodities in
the output (\code{TRUE}) or to remove them (\code{FALSE}, default). Useful
for debugging.}
}
\value{
The table provided in \code{input}, where the given column is
  replaced either by \code{faoID} or by \code{faoGroupID}, containing
  standardised commodity IDs.
}
\description{
This function matches commodities with a list of known commodities and
returns the matched IDs.
}
\details{
Arguments in \code{...} must be named either with \code{faoName},
  \code{faoGroupName} or \code{scientificName}, to match with the respective
  column in \code{id_commodities.csv}.
}
