% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchOntology.R
\name{matchOntology}
\alias{matchOntology}
\title{Match territories in a table with an ontology/gazetteer}
\usage{
matchOntology(
  table = NULL,
  columns = NULL,
  dataseries = NULL,
  ontology = NULL,
  verbose = FALSE,
  all_cols = FALSE
)
}
\arguments{
\item{table}{[\code{character(1)}]\cr a table that contains columns that
should be harmonised by matching with the gazetteer.}

\item{columns}{[\code{character(1)}]\cr the columns containing the concepts}

\item{dataseries}{[\code{character(1)}]\cr the source dataseries from which
territories are sourced.}

\item{ontology}{[\code{onto}]\cr either a path where the ontology/gazetteer
is stored, or an already loaded ontology.}

\item{verbose}{[`logical(1)`][logical]\cr whether or not to give detailed
information on the process of this function.}

\item{all_cols}{[`logical(1)`][logical]\cr whether or not to output all
tentative columns (for debugging), or only the essential columns.}
}
\description{
This function takes a table to replace the values of various columns with
harmonised values listed in the project specific gazetteer.
}
