% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchVars.R
\name{matchVars}
\alias{matchVars}
\title{Determine the valid ID of variables}
\usage{
matchVars(input = NULL, source = NULL, ..., keepOrig = FALSE)
}
\arguments{
\item{input}{[\code{character(.)}]\cr terms to be translated.}

\item{source}{[\code{integerish(1)}]\cr the census ID (\code{cenID}) from
which the terms have been taken.}

\item{...}{[\code{list(1)}]\cr lists that capture the variables by which to
match and the new column names containing the resulting ID; see Details.}

\item{keepOrig}{[\code{logical(1)}]\cr to keep the original commodities in
the output (\code{TRUE}) or to remove them (\code{FALSE}, default). Useful
for debugging.}
}
\value{
The table provided in \code{input}, where the given variable is
  replaced by the column that is specified by the argument name.
}
\description{
This function matches the values of a variable with an index and returns the
specified IDs.
}
\details{
Arguments in \code{...} are named lists, where the name of the list
  represents at the same time the variable to match and the the index in
  which to match (should have been created with \code{\link{makeIndex}}). The
  value represents the column in the index, with which to match the variable.
  The argument name represents the column in the index, which should be
  returned and which contains the new ID.
}
\examples{
\dontrun{

# match the column 'commodities' in 'someInput' with the column 'simpleName'
# in the index 'commodities' and chose the column 'faoID' in the index.
matchVars(input = someInput, faoID = list(commodities = "simpleName"))

# alternatively, the column 'faoGroupID' could be used via
matchVars(input = someInput, faoGroupID = list(commodities = "simpleName"))
}
}
