% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchYears.R
\name{matchYears}
\alias{matchYears}
\title{Determine the valid year}
\usage{
matchYears(input = NULL, priority = "harvest", keepOrig = FALSE)
}
\arguments{
\item{input}{[\code{data.frame(1)}]\cr table in which to match administrative
units.}

\item{priority}{[\code{character(1)}]\cr to which year should a tow-year date
be assigned, either the \code{"sowing"} or the \code{"harvest"} year.}

\item{keepOrig}{[\code{logical(1)}]\cr to keep the original units in the
output (\code{TRUE}) or to remove them (\code{FALSE}, default). Useful for
debugging.}
}
\value{
The table provided in \code{input}, where the given years are
  replaced by the column \code{years}, which contains the year to which the
  census statistics should be assigned.
}
\description{
This function matches years known years and returns \code{input} containing
the matched \code{years}.
}
\details{
On the southern hemisphere a growth period can span two years
  because the plants are sawn in autumn and the harvest is brought in in
  spring. This often results in a two-year notation, such as 1992-1993. This
  function thus mostly determines the correct year to which the census should
  be assigned.
}
