% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalise.R
\name{normalise}
\alias{normalise}
\title{Normlise census tables or geometries}
\usage{
normalise(what, ..., keepOrig = TRUE, update = FALSE, verbose = TRUE)
}
\arguments{
\item{what}{[\code{character(1)}]\cr the data to normalise, either
\code{"census"} or \code{"geometries"}.}

\item{...}{[\code{}]\cr for \code{what = "geometries"} chose a subset of
administrative units by calling \code{nation}, \code{un_member},
\code{continent}, \code{region} or \code{subregion}.\cr for \code{what =
"census"} chose}

\item{update}{[\code{logical(1)}]\cr whether or not the physical files should
be updated (\code{TRUE}) or the function should merely return the new
object (\code{FALSE} default).}

\item{verbose}{[\code{logical(1)}]\cr be verbose about what is happening
(default \code{TRUE}).}
}
\description{
Normlise census tables or geometries
}
