% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normTable.R
\name{normTable}
\alias{normTable}
\title{Normalise data tables}
\usage{
normTable(
  input = NULL,
  pattern = NULL,
  query = NULL,
  simplify = NULL,
  beep = NULL,
  verbose = FALSE
)
}
\arguments{
\item{input}{\code{\link[=character]{character(1)}}\cr path of the file to normalise. If
this is left empty, all files at stage two as subset by \code{pattern} are
chosen.}

\item{pattern}{\code{\link[=character]{character(1)}}\cr an optional regular expression.
Only dataset names which match the regular expression will be processed.}

\item{query}{\code{\link[=character]{character(1)}}\cr the expression that would be used
in \code{\link[dplyr]{filter}} to subset a tibble in terms of the columns
defined via the schema and given as a single character string, such as
\code{"ADM0 == 'a_nation'"}.}

\item{simplify}{\code{\link[=list]{list(.)}}\cr list of modules to simplify character
strings before matching with the ontology. Possible modules are: \itemize{
\item \code{"squish"}: logical of whether or not to replace duplicate whitespaces with single whitespaces,
\item \code{"lowercase"}: logical of whether or not to set all strings to lower case,
\item \code{"dashes"}: logical of whether or not to set all \href{https://en.wikipedia.org/wiki/Dash#Unicode}{dashes} and hyphens to \code{-} (unicode symbol U+002D),
\item \code{"duplpunct"}: logical of whether or not to remove duplicated punctation symbols (such as "- -" or \verb{//}),
\item \code{"remove"}: vector of symbols or regular expressions to remove,
\item \code{"replace"}: vector of two elements, the first contains a string to match, the second a string to replace the former with,
\item \code{"chartr"}: vector of two elements, the first contains n symbols to match, the second n symbols to replace the former with (see \link{chartr}).
}
These simplifications are aimed at reducing the need for manual
translations, for example, when it's known that the tables of a dataseries
have a lot of very similar symbols confused (for example en dash –, em
dash — or minus sign −) or systematic deviations (for example " - "
instead of merely "-") that could be rectified that way.}

\item{beep}{\code{\link[=integer]{integerish(1)}}\cr Number specifying what sound to be
played to signal the user that a point of interaction is reached by the
program, see \code{\link[beepr]{beep}}.}

\item{verbose}{\code{\link[=logical]{logical(1)}}\cr be verbose about translating terms
(default \code{FALSE}). Furthermore, you can use
\code{\link{suppressMessages}} to make this function completely silent.}
}
\value{
This function harmonises and integrates so far unprocessed data
tables at stage two into stage three of the areal database. It produces for
each main polygon (e.g. nation) in the registered data tables a file that
includes all thematic areal data.
}
\description{
Harmonise and integrate data tables into standardised format
}
\details{
To normalise data tables, this function proceeds as follows:
\enumerate{
\item Read in \code{input} and extract initial metadata from the file name.
\item Employ the function \code{\link[tabshiftr:reorganise]{tabshiftr::reorganise()}} to reshape \code{input} according to the respective schema description.
\item The territorial names are matched with the gazetteer to harmonise new territorial names (at this step, the function might ask the user to edit the file 'matching.csv' to align new names with already harmonised names). \item Harmonise territorial unit names.
\item store the processed data table at stage three.
}
}
\examples{
if(dev.interactive()){
  # build the example database
  adb_example(until = "normGeometry", path = tempdir())

  # normalise all available data tables ...
  normTable()

  # provide some options to simplify
  normTable(simplify = list(whitespace = TRUE,
                            replace = c("--", "-"),
                            chartr = c("æÜé", "aUe")))

  # ... and check the result
  output <- read_parquet(paste0(tempdir(), "/tables/stage3/a_nation.parquet"))
}
}
\seealso{
Other normalise functions: 
\code{\link{normGeometry}()},
\code{\link{normVocabulary}()}
}
\concept{normalise functions}
