% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regDataseries.R
\name{regDataseries}
\alias{regDataseries}
\title{Register a new geometry entry}
\usage{
regDataseries(name = NULL, description = NULL, website = NULL,
  notes = NULL, update = FALSE)
}
\arguments{
\item{name}{[\code{character(1)}]\cr the dataseries abbreviation.}

\item{description}{[\code{character(1)}]\cr the "long name" or "brief
description" of the dataseries.}

\item{website}{[\code{character(1)}]\cr the website, where the dataseries was
found or where additional information can be found.}

\item{notes}{[\code{character(1)}]\cr optional notes.}

\item{update}{[\code{logical(1)}]\cr whether or not the file
'id_dataseries.csv' should be updated.}

\item{address}{[\code{character(1)}]\cr an optional address of the dataseries
provider.}
}
\description{
Register a new geometry entry
}
\examples{
\dontrun{

setPath(root = "/home/se87kuhe/Nextcloud/LUCKINet/data/")

regDataseries(name = "maia",
              description = "Ministerio de Agroindustria",
              website = "http://datosestimaciones.magyp.gob.ar/",
              update = FALSE)
}
}
