% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorganise.R
\name{reorganise}
\alias{reorganise}
\title{Rectangrise a table}
\usage{
reorganise(input = NULL, schema = NULL)
}
\arguments{
\item{input}{[\code{data.frame(1)}]\cr table to rectangularise.}

\item{schema}{[\code{symbol(1)}]\cr the schema description for reorganising
\code{input}.}
}
\description{
This function takes a disorganised messy table and rearranges columns and
rows into a tidy dataset that can be sorted into the areal database.
}
