% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restoreIndex.R
\name{restoreIndex}
\alias{restoreIndex}
\title{Restore an index}
\usage{
restoreIndex(pattern = NULL, print = FALSE, select = NULL)
}
\arguments{
\item{pattern}{[\code{character(1)}]\cr name of the index that shall be
restored.}

\item{print}{[\code{logical(1)}]\cr whether or not a summary of the backed up
indices should be printed.}

\item{select}{[\code{character(1)}]\cr name of the table to which the table
should be reverted. Alternatively \code{"oldest"} or \code{"newest"}.}
}
\description{
Restore an index
}
\details{
Whenever an index is updated with \code{\link{updateTable}}, a
  backup of it is stored. When restoring one of these backup files with
  \code{restoreIndex}, the changes of \code{updateTable} will be rolled back
  to the status of the selected backup.
}
\examples{
\dontrun{

# first print the existing logs to chose from
restoreIndex(print = TRUE)

# then restore one
restoreIndex(pattern = "inv_geometries", select = "newest")
}
}
