% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_path.R
\name{select_path}
\alias{select_path}
\title{Select a system specific path}
\usage{
select_path(..., default = NULL)
}
\arguments{
\item{...}{combination of a nodename and the path that should be valid here.}

\item{default}{[`character(1)`][character]\cr a fallback path that should be
used in case the matching nodename is not mounted.}
}
\description{
This function calls \code{\link{Sys.info}} and selects from the provided
arguments that path that corresponds to the recent nodename.
}
\examples{
\dontrun{
myPath <- select_path(local = "/path/on/local/machine/",
                      hpc = "/path/on/remote/machine/",
                      default = "/fallback/path/if/nothing/else/found/")
}
}
