% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setIndex.R
\name{setIndex}
\alias{setIndex}
\title{Make and set an index}
\usage{
setIndex(input = NULL, output = NULL, origin = NULL, target = NULL,
  source = NULL)
}
\arguments{
\item{input}{[\code{tibble(1)}]\cr a possibly already existing table based on
which a translation table should be constructed.}

\item{output}{[\code{character(1)}]\cr name of the output file.}

\item{origin}{[\code{character(1)}]\cr column in \code{input} that contains the terms
that should be mapped to a translation.}

\item{target}{[\code{character(1)}]\cr column in \code{input} that contains the
standadrised terms to which \code{origin} should be mapped.}

\item{source}{[\code{character(1)}]\cr column in \code{input} that contains the
source/provenance of the table.}
}
\description{
Turn any table into a translation table.
}
\details{
To translate variables or territorial unit names one could use
  pre-existing translation tables, such as
  \href{https://en.wikipedia.org/wiki/Gazetteer}{gazetteers}. At the start of
  any project you would read in the translation tables that are relevant for
  your project, in case such tables do already exist. A table should be
  created for each variable that occurrs in your project, such as a table for
  territorial units, for species or for agricultural commodities.

  However, often a translation table doesn't contain all required mappings,
  or does not yet exist. Even if the table for a particular variable does not
  yet exist, you should create an empty table that could capture upcoming
  terms of the respective variable. When terms of a variable are encountered
  that can't be matched in any of the provided translation tables, the
  function asks you to translate them by hand. This is based on the functions
  \code{\link{translateTerms}} and \code{\link{updateIndex}}, however, those
  functions don't need to be called manually, they are instead called
  automatically, when unknown terms occur.
}
\examples{
library(readr)
setPath(root = "/home/se87kuhe/Nextcloud/LUCKINet/data/")

# create index from an already existing table
read_csv(file = "species.csv") \%>\%
   makeIndex(output = "mySpecies", origin = messySynonyms,
             target = scientificName)

# create empty table
makeIndex(output = "exoticVariable")
}
