% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setPath.R
\name{setPath}
\alias{setPath}
\title{Set the root path}
\usage{
setPath(root = NULL)
}
\arguments{
\item{root}{[\code{character(1)}]\cr path tothe root directory that contains
all tables and sub-directories.}
}
\description{
Gather and load all look-up-tables for translating and matching terms.
}
\examples{
setPath(root = "/home/se87kuhe/Nextcloud/LUCKINet/data/")
}
