% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setVariables.R
\name{setVariables}
\alias{setVariables}
\title{Make and set index and translation tables}
\usage{
setVariables(input = NULL, variable = NULL, type = "both",
  pid = NULL, origin = NULL, target = NULL)
}
\arguments{
\item{input}{[\code{tibble(1)}]\cr a possibly already existing table based on
which the output should be constructed.}

\item{variable}{[\code{character(1)}]\cr name of the variable and output file(s).}

\item{type}{[\code{character(1)}]\cr the type of table to create, either an
index (\code{"id"}), a translation table (\code{"tt"}) or \code{"both"}
(default).}

\item{pid}{[\code{character(1)}]\cr column in \code{input} that contains the
primary ID for the index table. If this is not given, an ID name is created
as \code{paste0(str_sub(variable, 1, 3), "ID")}}

\item{origin}{[\code{character(1)}]\cr column in \code{input} that contains
the terms that should be mapped to a translation.}

\item{target}{[\code{character(1)}]\cr column in \code{input} that contains
the standadrised terms to which \code{origin} should be mapped.}

\item{source}{[\code{character(1)}]\cr column in \code{input} that contains
the source/provenance of the table.}
}
\description{
Turn a table into a index and translation table.
}
\details{
To translate variables or territorial unit names one could use
  pre-existing translation tables, such as
  \href{https://en.wikipedia.org/wiki/Gazetteer}{gazetteers}. At the start of
  any project you would read in the translation tables that are relevant for
  your project, in case such tables do already exist. A table should be
  created for each variable that occurrs in your project, such as a table for
  territorial units, for species or for agricultural commodities.

  However, often a translation table doesn't contain all required mappings,
  or does not yet exist. Even if the table for a particular variable does not
  yet exist, you should create an empty table that could capture upcoming
  terms of the respective variable. When terms of a variable are encountered
  that can't be matched in any of the provided translation tables, the
  function asks you to translate them by hand. This is based on the functions
  \code{\link{translateTerms}} and \code{\link{updateIndex}}, however, those
  functions don't need to be called manually, they are instead called
  automatically, when unknown terms occur.
}
\examples{
library(readr)
setPath(root = "/home/se87kuhe/Nextcloud/LUCKINet/data/")

# create index from an already existing table
read_csv(file = "species.csv") \%>\%
   setVariables(variable = "mySpecies", origin = messySynonyms,
            target = scientificName)

# create empty table
setVariables(variable = "exoticVariable")
}
