% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{territories}
\alias{territories}
\title{Example \code{gazetteer}}
\format{
object of class \code{onto} for the example territories used in
\code{\link{adb_example}}.
}
\usage{
territories
}
\description{
An ontology of territory names (gazetteer)
}
\keyword{datasets}
