% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{testCompressed}
\alias{testCompressed}
\title{Test whether a file is a compressed file}
\usage{
testCompressed(x)
}
\arguments{
\item{x}{[\code{character(1)}]\cr the file name.}
}
\description{
(internal function not for user interaction)
}
\details{
This function looks at the file-extension and if it is one of
  \code{.gz}, \code{.bz2}, \code{.tar} \code{.zip}, \code{.tgz}, \code{.gzip}
  or \code{.7z}, it returns the value \code{TRUE}.
}
