% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateCensus.R
\name{updateCensus}
\alias{updateCensus}
\title{Update the census data}
\usage{
updateCensus(census = NULL, nations = NULL, file = NULL)
}
\arguments{
\item{census}{[\code{tibble(1)}]\cr the table that contains the standardised
census data.}

\item{nations}{[\code{character(1)}]\cr the nation name, which is also the
new file name.}

\item{file}{[\code{character(1)}]\cr full path of the stage2 file containing
census stats.}
}
\description{
Update the census data
}
