% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateData.R
\name{updateData}
\alias{updateData}
\title{Update the database}
\usage{
updateData(table = NULL, nations = NULL, file = NULL)
}
\arguments{
\item{table}{[\code{tibble(1)}]\cr the object that contains the standardised
data table.}

\item{nations}{[\code{character(1)}]\cr the nation name, which is also the
new file name.}

\item{file}{[\code{character(1)}]\cr full path of the stage2 file containing
stats.}
}
\description{
Store a harmonised and reshaped table into the database.
}
