% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateIndex.R
\name{updateIndex}
\alias{updateIndex}
\title{Update an index}
\usage{
updateIndex(index = NULL, name = NULL)
}
\arguments{
\item{index}{[\code{tibble(1)}]\cr the index to save.}

\item{name}{[\code{character(1)}]\cr name of the index that shall be updated.}
}
\description{
Update an index
}
