% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adb_backup.R
\name{adb_backup}
\alias{adb_backup}
\title{Backup the current state of an areal database}
\usage{
adb_backup()
}
\value{
No return value, called for the side effect of saving the inventory,
the stage3 files and modified ontology/gazetteer into the backup directory.
}
\description{
Backup the current state of an areal database
}
\details{
This function creates a tag that is composed of the version and the
date, appends it to all stage3 files (tables and geometries), the inventory
and the ontology/gazetteer files and stores them in the backup folder of
the current areal database.
}
