% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adb_diagnose.R
\name{adb_diagnose}
\alias{adb_diagnose}
\title{Diagnose databse contents}
\usage{
adb_diagnose(
  territory = NULL,
  concept = NULL,
  variable = NULL,
  level = NULL,
  year = NULL
)
}
\arguments{
\item{territory}{description}

\item{concept}{description}

\item{variable}{description}

\item{level}{description}

\item{year}{description}
}
\description{
work in progress, not yet useable
}
