% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adb_init.R
\name{adb_init}
\alias{adb_init}
\title{Initiate an areal database}
\usage{
adb_init(root, version, licence, gazetteer, top, ontology, variables)
}
\arguments{
\item{root}{[\code{character(1)}]\cr path to the root directory that contains
or shall contain an areal database.}

\item{version}{[\code{character(1)}]\cr version identifier for this areal
database.}

\item{licence}{[\code{character(1)}]\cr licence (link) for this areal
database.}

\item{gazetteer}{[\code{character(1)}]\cr path to the gazetteer that holds
the (hierarchical) information of territorial units used in this database.}

\item{top}{[\code{character(1)}]\cr the label of the class in the gazetteer
that represents the top-most unit (e.g. country) of the areal database that
shall be started.}

\item{ontology}{[\code{list(.)}]\cr named list with the path(s) of
ontologies, where the list name identifies the variable that shall be
matched with the ontology at the path.}

\item{variables}{[\code{character(.)}]\cr a set of variable names that are
available in this areal database.}
}
\value{
No return value, called for the side effect of creating the
  directory structure of the new areal database and tables that contain the
  database metadata.
}
\description{
Initiate a geospatial database or register a database that exists at the root
path.
}
\details{
This is the first function that is run in a project, as it initiates
  the areal database by creating the default sub-directories and initial
  inventory tables. When a database has already been set up, this function is
  used to register that path in the options of the current R session.
}
\examples{
adb_init(root = paste0(tempdir(), "/newDB"),
         version = "1.0.0", licence = "CC-BY-0.4",
         gazetteer = paste0(tempdir(), "/newDB/territories.rds"),
         top = "al1",
         ontology = list(var = paste0(tempdir(), "/newDB/ontology.rds")),
         variables = "test")

getOption("adb_path"); getOption("gazetteer_path")
}
