% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adb_query.R
\name{adb_query}
\alias{adb_query}
\title{Query data from the database}
\usage{
adb_query(..., variable = NULL, inventory = NULL, spatial = FALSE)
}
\arguments{
\item{...}{description}

\item{variable}{description}

\item{inventory}{[character(.)][character] one of \code{"dataseries"},
\code{"geometries"} or \code{"tables"}.}

\item{spatial}{description}
}
\value{
description
}
\description{
Query data from the database
}
\details{
Additional details...
}
\examples{
# example code
}
\concept{family name}
