% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adb_reset.R
\name{adb_reset}
\alias{adb_reset}
\title{Reset an areal database to its unfilled state}
\usage{
adb_reset(what = "all")
}
\arguments{
\item{what}{\code{\link[=logical]{logical(1)}}\cr what to reset, either \code{"onto"},
\code{"gaz"}, \code{"schemas"}, \code{"tables"}, \code{"geometries"} or
\code{"all"}, the default.}
}
\value{
no return value, called for its side effect of reorganising an areal
database into a state where no reg* or norm* functions have been run
}
\description{
Reset an areal database to its unfilled state
}
