% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adb_reset.R
\name{adb_reset}
\alias{adb_reset}
\title{Reset an areal database to its unfilled state}
\usage{
adb_reset(
  onto = TRUE,
  gaz = TRUE,
  schemas = TRUE,
  tables = TRUE,
  geometries = TRUE
)
}
\arguments{
\item{onto}{[\code{logical(1)}]\cr whether or not to reset the ontology.}

\item{gaz}{[\code{logical(1)}]\cr whether or not to reset the gazetteer}

\item{schemas}{[\code{logical(1)}]\cr whether or not to reset schemas.}

\item{tables}{[\code{logical(1)}]\cr whether or not to reset the tables.}

\item{geometries}{[\code{logical(1)}]\cr whether or not to reset the
geometries.}
}
\value{
no return value, called for its side effect of reorganising an areal
  database into a state where no reg* or norm* functions have been run
}
\description{
Reset an areal database to its unfilled state
}
