% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adb_translations.R
\name{adb_translations}
\alias{adb_translations}
\title{Load the translation tables of the currently active areal database}
\usage{
adb_translations(type = NULL, dataseries = NULL)
}
\arguments{
\item{type}{\code{\link[=character]{character(1)}}\cr the type of ontology for which to
load translation tables, either \code{"ontology"} to get the thematic
concepts, or \code{"gazetteer"} to get the territories.}

\item{dataseries}{\code{\link[=character]{character(1)}}\cr the name of a dataseries as
registered in \code{\link{regDataseries}}.}
}
\value{
returns the selected translation table
}
\description{
Load the translation tables of the currently active areal database
}
