% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adb_visualise.R
\name{adb_visualise}
\alias{adb_visualise}
\title{Visualise database contents}
\usage{
adb_visualise(
  ...,
  territory = NULL,
  concept = NULL,
  variable = NULL,
  level = NULL,
  year = NULL,
  diagnose = FALSE
)
}
\arguments{
\item{...}{combination of column name in the ontology and value to filter
that column by to build a tree of the concepts nested into it; see
\code{\link[ontologics]{make_tree}}.}

\item{territory}{description}

\item{concept}{description}

\item{level}{description}

\item{year}{description}

\item{diagnose}{description}
}
\value{
returns ...
}
\description{
Visualise database contents
}
