% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_variable.R
\name{get_variable}
\alias{get_variable}
\title{Extract concepts in stage 2 tables}
\usage{
get_variable(variable = NULL, dataseries = NULL)
}
\arguments{
\item{variable}{[\code{character(.)}]\cr name of the variable for which to
extract concepts.}

\item{dataseries}{[\code{integerish(1)}]\cr name of the dataseries from which
to extract the variables' concepts.}
}
\value{
the unique values at stage 2 of the specified dataseries
}
\description{
Tables at stage 2 are not yet harmonised and thus there is no unified way of
extracting information (this is only given at stage 3). In case these
information are required before the tables are harmonised, this function
allows to extract them.
}
\details{
This function pulls the schema description of all files of the given
  dataseries to harmonise the table and extact the \code{variable} values.
}
