% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchAdminUnits.R
\name{matchUnits}
\alias{matchUnits}
\title{Determine the valid hierarchy ID of administrative units}
\usage{
matchUnits(input = NULL, ..., keepOrig = FALSE)
}
\arguments{
\item{input}{[\code{data.frame(1)}]\cr table in which to match administrative
units.}

\item{...}{[\code{character(.)}]\cr define columns in \code{input}, where the
administrative units are recorded. Specify as \code{al1 = XYZ} to subset
administrative level 1 (nation) to "XYZ", \code{al2 = ...} for the second
level, etc.}

\item{keepOrig}{[\code{logical(1)}]\cr to keep the original units in the
output (\code{TRUE}) or to remove them (\code{FALSE}, default). Useful for
debugging.}
}
\value{
The table provided in \code{input}, where the given columns are
  replaced by the column \code{ahID}, which contains the administrative
  hierarchy ID.
}
\description{
This function matches administrative units with a list of known units and
returns \code{input} containing the matched \code{ahID}.
}
\details{
\code{...} must contain at least the argument \code{al1 = ...} to
  subset for a certain nation.
}
