% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normGeometry.R
\name{normGeometry}
\alias{normGeometry}
\title{Register geometries}
\usage{
normGeometry(input, update = FALSE, ...)
}
\arguments{
\item{input}{[\code{character(1)}]\cr path of the file to register.}

\item{update}{[\code{logical(1)}]\cr whether or not the physical files should
be updated (\code{TRUE}) or the function should merely return the new
object (\code{FALSE} default).}

\item{...}{[\code{}]\cr chose a subset of administrative units by calling
\code{nation}, \code{un_member}, \code{continent}, \code{region} or
\code{subregion}.}
}
\description{
Harmonise and integrate  geometries in a standardised format
}
