% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regDataseries.R
\name{regDataseries}
\alias{regDataseries}
\title{Register a new dataseries}
\usage{
regDataseries(name = NULL, description = NULL, website = NULL,
  notes = NULL, update = FALSE)
}
\arguments{
\item{name}{[\code{character(1)}]\cr the dataseries abbreviation.}

\item{description}{[\code{character(1)}]\cr the "long name" or "brief
description" of the dataseries.}

\item{website}{[\code{character(1)}]\cr a website where the dataseries or
additional information can be found.}

\item{notes}{[\code{character(1)}]\cr optional notes.}

\item{update}{[\code{logical(1)}]\cr whether or not the file
'inv_dataseries.csv' should be updated.}
}
\value{
Returns the entry that is appended to 'inv_dataseries.csv' in case
  \code{update = TRUE}.
}
\description{
This function registers a new dataseries of both, geometries or areal data
into the geospatial database
}
\examples{
\dontrun{

regDataseries(name = "gadm",
              description = "Database of Global Administrative Areas",
              website = "https://gadm.org/index.html",
              update = TRUE)

regDataseries(name = "usda",
              description = "US Dept. of Agriculture",
              website = "https://www.nass.usda.gov/Quick_Stats/Lite/index.php",
              update = TRUE)
}
}
