% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regTable.R
\name{regTable}
\alias{regTable}
\title{Register a new census data table}
\usage{
regTable(nation = NULL, subset = NULL, dSeries = NULL,
  gSeries = NULL, level = NULL, variable = NULL, algo = NULL,
  begin = NULL, end = NULL, archive = NULL, notes = NULL,
  update = FALSE)
}
\arguments{
\item{nation}{[\code{character(1)}]\cr the nation for which the boundaries
are defined.}

\item{subset}{[\code{character(1)}]\cr optional argument to specify which
subset the file contains. This could be a subset of administrative units or
of commodities.}

\item{dSeries}{[\code{character(1)}]\cr the dataseries of the census data.}

\item{gSeries}{[\code{character(1)}]\cr optinally, if the geometry series
deviates from the data series with which the census data should be
registered.}

\item{level}{[\code{integerish(1)}]\cr the administrative level at which the
boundaries are recorded.}

\item{variable}{[\code{character(.)}]\cr the variables included in
the file.}

\item{algo}{[\code{character}]\cr the dataseries specific algorithm
associated to the specific format of this file.}

\item{begin}{[\code{integerish(1)}]\cr the first date from which on the
boundaries are valid.}

\item{end}{[\code{integerish(1)}]\cr the last date from which on the
boundaries are valid.}

\item{archive}{[\code{character(1)}]\cr the original file from which the
boundaries emerge.}

\item{notes}{[\code{character(1)}]\cr optional notes.}

\item{update}{[\code{logical(1)}]\cr whether or not the file
'inv_sourceData.csv' should be updated.}
}
\description{
This function creates the standardized name for a new census file and enters
the required information into all look-up tables.
}
\details{
When processing census data, carry out the following steps:
  \enumerate{ \item Determine the \code{nation}, \code{administrative level},
  the \code{subset} of the nation and the \code{dataseries} of the shapefiles
  and provide them as arguments to this function. \item Provide a
  \code{begin} and \code{end} date for the census data. \item Run the
  function. \item (Re)Save the table with the following properties:
  \itemize{\item Format: csv \item Encoding: UTF-8 \item File name: What is
  provided as message by this function \item make sure that no columns or
  rows are removed. This will happen later via a \code{rect*} function that
  expects the original table.} \item Confirm that you have saved the file.}
}
\examples{
\dontrun{

setPath(root = "/home/se87kuhe/Nextcloud/LUCKINet/data/")

# dry run to be able to check whether everything is as intended.
regTable(nation = "Argentina", level = 3, dSeries = "maia",
         variable = c("planted_area", "harvested_area", "production", "yield"),
         algo = 1, begin = 1969, end = 2017,
         archive = "AgroIndustria_Estimaciones_complete_2017_11_12.xlsx")

# with several countries in one table (nation is NULL)
regTable(nation = NULL, level = 1, dSeries = "faostat", gSeries = "gadm",
         variable = c("planted_area","production","yield"),
         algo = 1, begin = 1961, end = 2017,
         archive = "Production_Crops_E_All_Data_(Normalized).csv")

# eventually, carry out the registration
regTable(nation = "Argentina", level = 3, dSeries = "maia",
         variable = c("planted_area", "harvested_area", "production", "yield"),
         algo = 1, begin = 1969, end = 2017,
         archive = "AgroIndustria_Estimaciones_complete_2017_11_12.xlsx",
         update = TRUE)
}
# when the data series and the geometry series are different, specify
# both of them
}
