% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translateTerms.R
\name{translateTerms}
\alias{translateTerms}
\title{Translate terms}
\usage{
translateTerms(terms, index = NULL, fuzzy_terms = NULL,
  fuzzy_dist = 5, verbose = TRUE)
}
\arguments{
\item{terms}{[\code{character(.)}]\cr terms to be translated.}

\item{index}{[\code{data.frame(1)}]\cr table that contains translations.}

\item{fuzzy_terms}{[\code{vector(.)}]\cr target terms with which a fuzzy
match should be carried out.}

\item{fuzzy_dist}{[\code{integerish(1)}]\cr the maximum distance for
fuzzy-matching.}

\item{verbose}{[\code{logical(1)}]\cr be verbose about what is happening
(default \code{TRUE}).}
}
\description{
Translate terms based on fuzzy matching.
}
\details{
This is basically a sophisticated matching algorithm, that adds new
  entries to the respective index, if no match was found.
}
