% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unifyNations.R
\name{unifyNations}
\alias{unifyNations}
\title{Unify the names of Nations}
\usage{
unifyNations(unify = NULL, nameCol = NULL)
}
\arguments{
\item{unify}{[\code{data.frame(1)}]\cr the object in which to unify nations.}

\item{nameCol}{[\code{character}]\cr the name of the column that contains
nation names to unify.}
}
\description{
In LUCKINet we use a default set of nations. This functions renames nations
that don't abide by those rules yet.
}
\examples{
\dontrun{

library(sf)

setPath(root = "/home/se87kuhe/Nextcloud/LUCKINet/data/")
gadm <- st_read(paste0(getOption("dmt_path"),
                       "administrative_boundaries/original_datasets/gadm36_levels.gpkg"),
               layer = "level0",
               stringsAsFactors = FALSE)

(gadm2 <- unifyNations(unify = gadm, nameCol = "NAME_0"))
}
}
