% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unifyNations.R
\name{unifyNations}
\alias{unifyNations}
\title{Unify the names of Nations}
\usage{
unifyNations(unify = NULL, source = NULL, nameCol = NULL,
  verbose = TRUE)
}
\arguments{
\item{unify}{[\code{data.frame(1)}]\cr the object in which to unify nations.}

\item{source}{[\code{integerish(1)}]\cr and ID, typically the census ID from
which the terms are provided.}

\item{nameCol}{[\code{character}]\cr the name of the column that contains
nation names to unify.}

\item{verbose}{[\code{logical(1)}]\cr be verbose about what is happening
(default \code{TRUE}).}
}
\description{
In LUCKINet we use a default set of nations. This functions renames nations
that don't abide by those rules yet.
}
\examples{
\dontrun{

library(sf)

setPath(root = "/home/se87kuhe/Nextcloud/LUCKINet/data/")
gadm <- st_read(paste0(getOption("cT_path"),
                       "ct_geometries/stage1/gadm36_levels.gpkg"),
               layer = "level0",
               stringsAsFactors = FALSE)

(gadm2 <- unifyNations(unify = gadm, nameCol = "NAME_0"))
}
}
