% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateTable.R
\name{updateTable}
\alias{updateTable}
\title{Update a table}
\usage{
updateTable(index = NULL, name = NULL, matchCols = NULL, backup = TRUE)
}
\arguments{
\item{index}{[\code{tibble(1)}]\cr the table to use as update.}

\item{name}{[\code{character(1)}]\cr name of the table that shall be updated.}

\item{matchCols}{[\code{character(.)}]\cr the columns in the old file by
which to match.}

\item{backup}{[logical(1)]\cr whether or not to store the old table in the
\code{log} directory.}
}
\description{
Update any inventory, index or translation table of an areal database.
}
