% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateTable.R
\name{updateTable}
\alias{updateTable}
\title{Update a table}
\usage{
updateTable(index = NULL, name = NULL)
}
\arguments{
\item{index}{[\code{tibble(1)}]\cr the table to use as update.}

\item{name}{[\code{character(1)}]\cr name of the table that shall be updated.}
}
\description{
Update any inventory, index or translation table of a geospatial
database.
}
